refactor dado primitivo de double para BigDecimal

+++ src/main/java/br/com/naturaves/cobrancanaturaves/boleto/application/api/BoletoClienteListResponse.java
@@ -1,5 +1,6 @@
 package br.com.naturaves.cobrancanaturaves.boleto.application.api;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -15,7 +16,7 @@ public class BoletoClienteListResponse {
 	private String documento;
 	private String numeroDaParcela;
 	private LocalDate dataVencimento;
-	private Double saldoDevedor;
+	private BigDecimal saldoDevedor;
 	private GrupoEmpresarial grupoEmpresarial;
 	private LocalDateTime dataHoraDaCadastro;
